refactor/replace component and rename files
