+++ src/features/tutorApplication/index.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
 
-import Button from '../../../shared/ui/button';
-import Stepper from '../../../shared/ui/stepper';
-import SubjectForm from './subject-form';
+import Button from '../../shared/ui/button';
+import Stepper from '../../shared/ui/stepper';
+import SubjectForm from './ui/subjectForm';
 
 import styles from './index.module.scss';
