+++ src/features/tutorApplication/ui/ProfileInfo/AvatarBlock/index.module.scss
