+++ src/features/tutorApplication/ui/ProfileInfo/AvatarBlock/index.tsx
