+++ src/features/tutorApplication/ui/ProfileInfo/AvatarBlock/type.ts
