+++ src/features/tutorApplication/ui/ProfileInfo/AvatarUploadModal/index.module.scss
