+++ src/features/tutorApplication/ui/subjectForm/index.tsx
@@ -2,7 +2,7 @@ import { useState } from 'react';
 import React from 'react';
 
 import iconAdd from '../../../../assets/icons/icon_add.svg';
-import SubjectFormItem from './subject-form-item';
+import SubjectFormItem from './subjectFormItem';
 
 import styles from './index.module.scss';
