+++ src/features/tutorApplication/ui/subjectForm/subjectFormItem/profileCategories/index.module.scss
